Add files via upload
This is a big side project I've been working on, and I may refactor the entire thing soon, because I hard-coded too much and now it is a mess. The idea is a text-based adventure game where the player can navigate around a world, interact with objects, and solve puzzles to progress. My hope is to write it much more abstracted and with a coordinates system to keep track of everything. I continue to work on it in my free time.